+++ pm4py/algo/clustering/profiles/__init__.py
@@ -1,0 +1,1 @@
+from pm4py.algo.clustering.profiles import algorithm, variants