+++ pm4py/algo/clustering/profiles/variants/__init__.py
@@ -1,0 +1,1 @@
+from pm4py.algo.clustering.profiles.variants import sklearn_profiles