+++ pm4py/objects/ocel/exporter/xmlocel/exporter.py
@@ -17,13 +17,14 @@
 from enum import Enum
 from typing import Optional, Dict, Any
 
-from pm4py.objects.ocel.exporter.xmlocel.variants import classic
+from pm4py.objects.ocel.exporter.xmlocel.variants import classic, ocel20
 from pm4py.objects.ocel.obj import OCEL
 from pm4py.util import exec_utils
 
 
 class Variants(Enum):
     CLASSIC = classic
+    OCEL20 = ocel20
 
 
 def apply(ocel: OCEL, target_path: str, variant=Variants.CLASSIC, parameters: Optional[Dict[Any, Any]] = None):